requirements.txt
setuptools >= 21.0.0
elasticsearch==7.13.0
fastapi==0.101.1
httpx==0.24.1
pytz==2023.3
pytest-cov==4.1.0
prometheus-client==0.16.0
prometheus-fastapi-instrumentator==6.0.0
databases==0.5.3
SQLAlchemy==1.4.26
fastapi-pagination==0.12.9
python-dotenv==1.0.0
uvicorn==0.23.2
gunicorn==21.2.0
psycopg2-binary==2.9.7
pika==1.3.2
aio-pika==9.2.3
anyio==4.0.0
aiormq==6.7.7
pamqp==3.2.1
redis==5.0.1
retry==0.9.2
langchain==0.0.310
beautifulsoup4==4.12.2
lxml==4.9.2
tiktoken==0.3.3